fix bug

+++ src/features/task/components/task-breadcrumbs.tsx
@@ -8,11 +8,13 @@ import { Button } from "@/components/ui/button";
 import { useConfirm } from "@/components/confirm";
 import { deleteTaskService } from "@/services/taskService";
 import { useRouter } from "next/navigation";
+import { WorkspaceMember } from "@/types/workspace";
 interface TaskBreadcrumbProps {
   project: ProjectResponse;
   task: ResponseTask;
+  currentMember: WorkspaceMember;
 }
-export const TaskBreadcrumb = ({ project, task }: TaskBreadcrumbProps) => {
+export const TaskBreadcrumb = ({ project, task,currentMember }: TaskBreadcrumbProps) => {
   const [ConfirmDialog, confirm] = useConfirm(
     "Delete task",
     "This action cannot be undone.",
@@ -21,9 +23,16 @@ export const TaskBreadcrumb = ({ project, task }: TaskBreadcrumbProps) => {
   const { mutate: deleteTaskMutate, isPending } = deleteTaskService();
   const router = useRouter();
 
+
   const onDelete = async () => {
+    if (!currentMember || currentMember.role !== "OWNER") {
+      
+      return alert("You do not have permission to delete this task.");
+    }
+  
     const ok = await confirm();
     if (!ok) return;
+  
     deleteTaskMutate(
       {
         workspaceId: task.workspaceId,
@@ -32,15 +41,17 @@ export const TaskBreadcrumb = ({ project, task }: TaskBreadcrumbProps) => {
       },
       {
         onSuccess: () => {
-         
-          router.push(`/workspaces/${task.workspaceId}/projects/${project.id}`);
+          router.push(
+            `/workspaces/${task.workspaceId}/projects/${project.id}`
+          );
         },
         onError: (error) => {
           console.error("Failed to delete task:", error);
         },
-      },
+      }
     );
   };
+  
   console.log(task.workspaceId, project.id, task.id);
   return (
     <div className="flex items-center gap-x-2">
